OCR/aadhaar_api/requirements.txt
fastapi
uvicorn
PyPDF2
pdf2image
pytesseract
opencv-python
Pillow